Add config provider & php backed config
