+++ src/ConfigProvider/ConfigProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tnapf\Config\ConfigProvider;
+
+interface ConfigProvider
+{
+    /**
+     * @throws InvalidConfigException
+     */
+    public function get(string $key): mixed;
+}
