+++ tests/E2E/config/invalid-config.php
